Make sure that the OWLAPI ontology object is loaded. Even if it is not, we continue to try to delete the dataset in Solr and Virtuoso in case that the ontology is not loaded, un-loadable, and that the other data instances needs to be cleaned-up.

+++ ontology/delete/OntologyDelete.php
@@ -582,9 +582,12 @@ class OntologyDelete extends WebService
       $this->initiateOwlBridgeSession();
 
       $this->getOntologyReference();
-      
+
       // Delete the OWLAPI instance
-      $this->ontology->delete();
+      if($this->ontology)
+      {
+        $this->ontology->delete();
+      }
       
       // Remove the holdOntology tag before deleting the ontology
       $query = "delete data from <" . $this->wsf_graph . "datasets/>
